simlify pkg test script; readme updates and add list of problems; pin bazel version; ensure that docker is used for MakeDeb and PackageTar actions
The MakeDeb and PackageTar actions were just being executed on the host with `--config docker`. Note that `remote` is the highest priority strategy enabled by default, so we've not had to override it as we do with the experimental docker sandbox.

CentOS6's Python 3.4 is too old and is not supported for pkg_tar (though is for pkg_rpm!) we could avoid this by enforing the MakrTar action to run not-remote, i.e.  "remote,worker,sandboxed,local" -> "worker,sandboxed,local" or possible by removing the exec_platforms.